case switch stuff
today I was exteremly productive as one can see by looking at the freaking amazing comit to github...in this comit we can see that I did some work on case and switches which is an infior version of if/else if statments which we all know are a supior version of case and switch statements.   Anywho my next comit wont be for a while because I wont be here next week and who knows about any week after that.  In that case it has been fun most of the time

cool, cool, cool, cool, cool.  No doubt, no doubt, no doubt